Move course cards to @yoast/components
