+++ apps/components/ComponentsExample.js
@@ -2,11 +2,9 @@ import React from "react";
 import styled from "styled-components";
 import { __ } from "@wordpress/i18n";
 
-import { FullHeightCard } from "yoast-components/composites/CoursesOverview/Card";
-import CardDetails from "yoast-components/composites/CoursesOverview/CardDetails";
+import { CardDetails, FullHeightCard, YoastWarning } from "@yoast/components";
 import getCourseFeed from "yoast-components/utils/getCourseFeed";
 import { getDirectionalStyle } from "@yoast/helpers";
-import { YoastWarning } from "@yoast/components";
 
 const Container = styled.div`
 	max-width: 1024px;
